requirements.txt
accelerate==1.2.1
optimum-quanto==0.2.6
peft==0.14.0
sentencepiece==0.2.0
transformers==4.44.2
diffusers @ git+https://github.com/huggingface/diffusers@dc62e6931e3aac1d78375928693186764dcb8492
datasets
pandas==2.2.3
gradio
pillow==11.1.0
protobuf
torch==2.4.1
torchvision
numpy==1.26.4
deepspeed==0.14.4
einops==0.8.0
huggingface-hub==0.24.5
pandas==2.2.2
opencv-python==4.10.0.84
supervision==0.25.0
cog
peft
pillow
requests
